统一uint

+++ src/model/comment.go
@@ -4,7 +4,7 @@ import "github.com/jinzhu/gorm"
 
 type Comment struct { // 评论
 	gorm.Model
-	VideoId int64  `json:"video_id,omitempty"`
-	UserId  int64  `json:"user_id,omitempty"`
+	VideoId uint   `json:"video_id,omitempty"`
+	UserId  uint   `json:"user_id,omitempty"`
 	Content string `json:"content,omitempty"`
 }
