+++ src/model/video.go
@@ -4,10 +4,10 @@ import "github.com/jinzhu/gorm"
 
 type Video struct {
 	gorm.Model
-	AuthorId      int32  `json:"author"`
+	AuthorId      uint   `json:"author"`
 	PlayUrl       string `json:"play_url"`
 	CoverUrl      string `json:"cover_url"`
-	FavoriteCount int64  `json:"favorite_count"`
-	CommentCount  int64  `json:"comment_count"`
+	FavoriteCount uint   `json:"favorite_count"`
+	CommentCount  uint   `json:"comment_count"`
 	Title         string `json:"title"`
 }
